requirements.txt
fastapi==0.115.14
uvicorn[standard]==0.35.0
pydantic==2.12.4
python-dotenv==1.1.1
requests==2.32.5
pytest==8.4.2
